Udated to move Navigation code to separate JS file
